feat: merge l1en into dev

+++ src/contract/v3/abi/bonding-curve.ts
@@ -394,14 +394,14 @@ export const v3BondingCurveAbi = [
             type: 'bool',
           },
           {
-            internalType: 'uint8',
+            internalType: 'uint40',
             name: 'distributionRatioKol',
-            type: 'uint8',
+            type: 'uint40',
           },
           {
-            internalType: 'uint8',
+            internalType: 'uint40',
             name: 'distributionRatioCommunity',
-            type: 'uint8',
+            type: 'uint40',
           },
           {
             internalType: 'uint40',
